Move MyTripsListView to organisms/ from molecules/
